Abandon trie<T>::iterator's const-correctness
The whole concept is *this* close to working fine.  But GCC 4.8.x
does not support the prototype

  iterator std::vector::erase(const_iterator)

which trie<T>::erase would need to use if trie<T>::iterator used a
const_iterator over a trie<T>'s children vector (which it should--
there is no need to modify the children.)

I could reinterpret_cast<> the const_iterator and hope it worked,
but that's kludgier than removing const-correctness.  Once GCC 4.9
is released, please, rever this commit.